requirements.txt
# Backend requirements
Flask==2.3.3
pytesseract==0.3.10
pdf2image==1.16.3
Pillow==9.4.0
spacy==3.5.0
PyPDF2

# Frontend requirements
streamlit==1.26.0
requests==2.31.0

# Zusätzliche Abhängigkeiten für spacy (falls erforderlich)
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.5.0/en_core_web_sm-3.5.0-py3-none-any.whl
